% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_figure_vast.R
\name{add_figure_vast}
\alias{add_figure_vast}
\title{Write a csv file of figures for a given VAST index}
\usage{
add_figure_vast(dir, outfile)
}
\arguments{
\item{dir}{The directory you want to search for VAST files in.}
}
\description{
Find four files of interest written to the disk from VASTWestCoast
and include their details in a csv file that is helpful for
including figures in main document.
}
\seealso{
\link[sa4ss:add_figure]{sa4ss::add_figure}; \link{add_figure_ling}
}
