% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_control_fleets.R
\name{change_control_fleets}
\alias{change_control_fleets}
\title{renumber fleets in lingcod control file}
\usage{
change_control_fleets(area, fleets, ctl)
}
\arguments{
\item{area}{area, initially either "n" or "s" for 2021 models}

\item{fleets}{table of info on old and new fleets created by get_fleets()}

\item{ctl}{control file object created by \code{\link[=get_inputs_ling]{get_inputs_ling()}} or
\code{\link[r4ss:SS_readctl]{r4ss::SS_readctl()}}}
}
\description{
function hopefully only to be used once for each model
}
\seealso{
\code{\link[=change_data_fleets]{change_data_fleets()}}, \code{\link[=get_fleet]{get_fleet()}}, \code{\link[=get_inputs_ling]{get_inputs_ling()}}
}
\author{
Ian G. Taylor
}
