% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_data_fleets.R
\name{change_data_fleets}
\alias{change_data_fleets}
\title{renumber fleets in lingcod data file}
\usage{
change_data_fleets(area, fleets, dat)
}
\arguments{
\item{area}{area, initially either "n" or "s" for 2021 models}

\item{fleets}{table of info on old and new fleets created by get_fleets()}

\item{dat}{data file object created by get_inputs_ling()}
}
\description{
function hopefully only to be used once for each model
}
\seealso{
\code{\link[=change_control_fleets]{change_control_fleets()}}, \code{\link[=get_fleets]{get_fleets()}}, \code{\link[=get_inputs_ling]{get_inputs_ling()}}
}
\author{
Ian G. Taylor
}
