% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_comps.R
\name{clean_comps}
\alias{clean_comps}
\title{change the column names in the data.frame of length or age comps}
\usage{
clean_comps(comp, type = "len")
}
\arguments{
\item{comp}{data object in the lingcod package}

\item{verbose}{A logical value specifying if output should be printed
to the console or not. The default is \code{FALSE}, which will suppress messages.}
}
\description{
aligns the  names from the various software tools and combines
data frames with sexed and unsexed comps into one
}
\seealso{
\code{\link[=add_data]{add_data()}}
}
\author{
Ian G. Taylor
}
