% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_year.R
\name{coef_year}
\alias{coef_year}
\title{Get model coefficients for year}
\usage{
coef_year(results)
}
\arguments{
\item{results}{A model object output from a statistical model
such as \link[stats:glm]{stats::glm}.}
}
\value{
A vector of parameter values for all parameters pertaining
to year in the model.
}
\description{
Call \link[stats:coef]{stats::coef} to get model coefficients from a model fit
in R. Year will be a factor term, and thus, the coefficients with
year in their name will be returned. No interactions with year are
included.
}
\author{
Kelli F. Johnson
}
