% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_ling.R
\name{compile_ling}
\alias{compile_ling}
\title{Compile the stock assessment document(s) for a given species}
\usage{
compile_ling(dir = "North", wipe = FALSE, ...)
}
\arguments{
\item{dir}{A terribly named parameter that specifies which pdf you want to build.
For lingcod, this is just \code{"North"} or \code{"South"}.
Also for just lingcod, we can use \code{get_groups(info_groups)} to get a vector of
all the groups we have, which results in both documents being built.}

\item{wipe}{A logical values specifying if you want to remove all cached objects.
For lingcod, this will remove plots, custom_plots, and cached markdown files and
will lead to the document taking approximately an hour to build.}

\item{...}{Arguments you want to pass to \link[bookdown:render_book]{bookdown::render_book}.}
}
\description{
Compile the stock assessment document(s) for a given species where the number of
documents depends on the number of directories that you pass to dir.
Typically, this happens in the directory called \code{doc} that stores a single
directory for each stock.
\link[bookdown:render_book]{bookdown::render_book} will be called inside on each directory in \code{dir}.
}
\details{
status:
This function is currently in beta testing, which means that it
\itemize{
\item might be helpful,
\item could contain bugs,
\item will more than likely have non-descript error and warning messages.
Please post an \href{https::/github.com/nwfsc-assess/sa4ss/issues}{issue}
or email the author of this function.
Thank you for helping test it out, and we hope the pain is worth the gain!
}

index.Rmd
When specifying a directory to \link[bookdown:render_book]{bookdown::render_book}, the
file \code{index.Rmd} is normally used as the master file for rendering the book
and all other files are brought in as chapters in alphabetical order.
Here, in the \pkg{sa4ss} package we use \verb{00a.Rmd} as the indexing file and
all other .Rmd files are sourced in alphabetical order.
The renaming is done automatically for you in the call to \link{draft} and just
mentioned here for completeness.
}
\examples{
\dontrun{
# An example for lingcod in 2021.
compile_ling(dir = file.path(get_groups(info_groups)))
}

}
\seealso{
See \link{compile_internal} for the non-vectorized version of what happens
inside each directory.

Other compile: 
\code{\link{compile_changebookname}()},
\code{\link{compile_internal}()}
}
\author{
Kelli F. Johnson and undergoing testing by Kelli F. Johnson
}
\concept{compile}
