% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_output.R, R/tables_cpue.R
\name{cpue_output}
\alias{cpue_output}
\title{Read in OR index table information}
\usage{
cpue_output(grep, fleet)

cpue_output(grep, fleet)
}
\arguments{
\item{grep}{A regex for the file names in \code{data-raw}
that will pick up all three files.}

\item{fleet}{The fleet name from \link{get_fleet}.}
}
\description{
Read in OR index table information

Read in OR index table information
}
\examples{
unlink(file.path("tables", "tables_cpue.csv"))
unlink(file.path("figures", "figures_cpue.csv"))
cpue_output(
  grep = "OR_ORBS_CPUE", fleet = "OR"
)
cpue_output(
  grep = "OR_COMM_NSLOG_CPUE", fleet = "Fix"
)
\dontrun{
unlink(file.path(dirname(system.file(package=utils_name())),
  "tables", "tables_cpue.csv")
)
unlink(file.path(dirname(system.file(package=utils_name())),
  "figures", "figures_cpue.csv")
)
cpue_output(
  grep = "OR_COMM_NSLOG_CPUE", fleet = "Fix"
)
cpue_output(
  grep = "OR_ORBS_CPUE", fleet = "OR"
)
}
}
