% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_catch}
\alias{data_catch}
\title{Model catches}
\format{
A data frame with four columns
\itemize{
\item Year: a four digit integer for year
\item area: a character value denoting the area the catches pertain to
\item fleet: character values denoting the fleet, with
fixed gear (FG), trawl (TW), and abbreviations for state are all
valid entries and states signify recreational catches
\item mt: catches in metric tons (mt)
}
}
\usage{
data_catch
}
\description{
A data frame of commercial and recreational catches for
the northern and southern areas by fleet.
}
\keyword{datasets}
