% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_index_cpue}
\alias{data_index_cpue}
\title{Index for catch per unit effort (CPUE)}
\format{
A data frame with the following columns
\itemize{
\item year: a four digit integer for year
\item seas: equal to 7 because all fleets "operate" in the middle of the year
\item index: equal to the proper fleet number returned from
\link{get_fleet}
\item obs: the median of the posterior estimate per year
\item se_log: the log standard error of the estimate
\item area: e.g., north or south
\item source: which file the data came from because there are multiple
versions of the California recreational fleet
}
}
\usage{
data_index_cpue
}
\description{
A data frame formatted for SS with information on the
CPUE indices.
}
\keyword{datasets}
