% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_index_RecOR}
\alias{data_index_RecOR}
\title{Index for Oregon recreational fishery}
\format{
A data frame with four columns
\itemize{
\item year: a four digit integer for year
\item seas: equal to 7 because all fleets "operate" in the middle of the year
\item index: equal to the proper fleet number returned from
\link{get_fleet}
\item obs: the median of the posterior estimate per year
\item se_log: the log standard error of the estimate
}
}
\usage{
data_index_RecOR
}
\description{
A data frame formatted for SS with information on the
Oregon recreational fishery CPUE.
}
\keyword{datasets}
