% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_index_survey}
\alias{data_index_survey}
\title{Index for the standardized surveys}
\format{
A data frame with four columns
\itemize{
\item Year: a four digit integer for year
\item dir: the directory where the data are stored in data-raw
\item surveyname: the survey name using characters
\item area: the model area of relevance, e.g., North
\item distribution: the assumed distribution used to fit the model
\item Fleet: a version of the strata with depth
\item Estimate_metric_tons: estimate in mt
\item SD_log: the log standard deviation
\item SD_mt: the standard deviation in normal space
\item strata: the stratafication the estimate pertains to
\item year: a four digit integer for year
\item obs: the estimate per year
\item se_log: the log standard error of the estimate
\item seas: equal to 7 because all fleets "operate" in the middle of the year
\item index: equal to 7 because all fleets "operate" in the middle of the year
}
}
\usage{
data_index_survey
}
\description{
A data frame of index values for the standardized surveys that
are relevant to this species.
}
\keyword{datasets}
