% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_order.R
\name{dir_order}
\alias{dir_order}
\title{Order a vector of files based on their manipulation time.}
\usage{
dir_order(..., grep)
}
\arguments{
\item{...}{The creation of a file path using \code{...} where you can
provide a single value or you can give individual sections of a path.
This will be the directory that is searched.}

\item{grep}{A regular expression you want to search for in the
directory.}
}
\value{
A vector of file names that are in order from the time that
the files were manipulated.
}
\description{
Requires that you run someone's full script and not just update a single
figure because then the manipulation time will be off.
}
\examples{
dir_order("tables", grep = "-|_")
dir_order(c("tables", "figures"), grep = "cpuewa")
}
\author{
Kelli F. Johnson
}
