% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_rates_combined.R
\name{discard_rates_combined}
\alias{discard_rates_combined}
\title{Combine discard rates from multiple sources}
\usage{
discard_rates_combined(
  info_ncs,
  info_cs,
  info_em,
  Area = "North4010",
  gears = NULL,
  fleet = NA,
  min_cv = 0.05
)
}
\arguments{
\item{info_ncs}{Table for non-catch-shares discard info}

\item{info_cs}{Table for catch-shares discard info}

\item{info_em}{Table for electronic monitoring discard info}

\item{Area}{Area label in tables provided by Chantel}

\item{gears}{Subset of gears to include in the calculations}

\item{fleet}{Name or number of the fleet to use in the "Flt" column
in the table for use in SS}

\item{min_cv}{Minimum CV to use for fleets with 100\% monitoring and no bootstrap
uncertainty provided}
}
\value{
Table for input into SS data file for a specific fleet.
}
\description{
Combines values from the tables provided by Chantel Wetzel to
get a single fleet-specific discard rate after aggregating
across gears and sectors/observation method. This function needs to be
run separately for each fleet to be used in the model.
}
\author{
Ian G. Taylor
}
