% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_local.R
\name{file_local}
\alias{file_local}
\title{Find a file specific to the lingcod package}
\usage{
file_local(...)
}
\arguments{
\item{...}{Sections of a file path passed to \link{system.file}.
This will be valid for anything inside of \verb{inst\\}.}
}
\description{
A shortcut for \link{system.file} when you get tired of typing which package
you need to find the file in.
}
\examples{
# List the files in the directory inst\extdata
dir(file_local("extdata"))

}
\author{
Kelli F. Johnson thanks to \pkg{bookdown}.
}
