% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_range.R
\name{format_range}
\alias{format_range}
\title{Format text for a range from a vector}
\usage{
format_range(x, parentheses = TRUE)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{parentheses}{A logical value specifying if you want the value returned
to be wrapped in parentheses or not. By default, they are included.}
}
\value{
A formatted character string.
}
\description{
Format the range (i.e., minimum and maximum) from a vector
for inclusion in a text document with or without parentheses.
}
\details{
todo:
Format the range values to have a given number of digits if users are not
using it for just a year range.
}
\examples{
format_range(1:12)
}
\author{
Kelli Faye Johnson
}
