% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dir_exe.R
\name{get_dir_exe}
\alias{get_dir_exe}
\title{Get directory with Stock Synthesis executable in Lingcod package}
\usage{
get_dir_exe()
}
\description{
Should find executable within package regardless of the location
on any user's system. Could be expanded in the future for different
operating systems.
}
\examples{
\dontrun{
  # find directory with executable
  get_dir_exe()

  # copy model files to a new directory and bring in executable
  r4ss::copy_SS_inputs(dir.old = get_dir_ling(area = area, num = 1),
                       dir.new = get_dir_ling(area = area, num = 2), 
                       use_ss_new = FALSE,
                       copy_exe = TRUE,
                       dir.exe = get_dir_exe())
}
}
\seealso{
\code{\link[=get_dir_ling]{get_dir_ling()}}
}
\author{
Ian G. Taylor
}
