% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dir_ling.R
\name{get_dir_ling}
\alias{get_dir_ling}
\title{get directory for a lingcod model}
\usage{
get_dir_ling(
  area = NULL,
  num = NULL,
  sens = 1,
  yr = 2021,
  id = NULL,
  verbose = FALSE
)
}
\arguments{
\item{area}{area, initially either "n" or "s" for 2021 models}

\item{num}{base model number}

\item{sens}{sensitivity number (assumed 1 if not specified)}

\item{yr}{model year (assumed 2021 if not specified)}

\item{id}{model id (like "2021.s.001.001") as an alternative to
the combination of inputs \code{area}, \code{num}, \code{sens}, and \code{year}}

\item{verbose}{A logical value specifying if output should be printed
to the console or not. The default is \code{FALSE}, which will suppress messages.}
}
\description{
uses either a model id like "2021.s.001.001" or the combination of
year (\code{yr}), area (\code{area}), run number (\code{num}), and sensitivity number
(\code{sens}) to assemble that id, then returns the associated directory
based on the table in the models/README.md
}
\examples{
\dontrun{
  get_dir_ling(area ="s", num = 1)
  get_dir_ling(id = "2021.s.001.001")
}
}
\seealso{
\code{\link[=get_inputs_ling]{get_inputs_ling()}}
}
\author{
Ian G. Taylor
}
