% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fleet.R
\name{get_fleet}
\alias{get_fleet}
\title{translate between fleet names and numbers and old or new}
\usage{
get_fleet(value = NULL, yr = 2021, area = NULL, ignore.case = TRUE)
}
\arguments{
\item{value}{either a fleet number or a string matching the fleet name
a NULL value will return the full table of all fleet info for the associated
year (separate TRI_Early and TRI_Late will be returned with yr = 2019, not
yr = 2019).
a vector of numeric values will return a vector}

\item{yr}{year (assumed 2021 if not specified)}

\item{area}{either "n" or "s"}

\item{ignore.case}{if FALSE, the pattern matching is case sensitive and
if TRUE, case is ignored during matching (passed to \code{grep})}
}
\description{
uses models/fleets.csv as a lookup table to translate between
different combinations
}
\examples{
  # current fleet number for trawl fishery
  get_fleet("trawl")$num

  # plain text label associated with fleet 3
  get_fleet(3)$label_short

  # current fleet number that corresponds to an old fleet name
  get_fleet("TRI_Early", area = "s", yr = 2019)$num

  # current fleet number that corresponds to an old fleet number
  get_fleet(4, area = "s", yr = 2019)$num

  # current fleet number that corresponds to a vector of old fleet numbers
  get_fleet(value = c(1,2,2,3), yr = 2019, area = 's')$num

  # info on all current rec fleets
  get_fleet("rec")

  # info on all fleets in 2021 models (excludes duplicates)
  get_fleet()

  # info on all fleets in 2019 models
  # (includes separate rows for TRI_Early and TRI_Late)
  get_fleet(yr = 2019)
}
\author{
Ian G. Taylor
}
