% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_groups}
\title{Get grouping structure for a given species}
\usage{
get_groups(data, collapsecharacter = "-")
}
\arguments{
\item{data}{A data frame with a potentially infinite number of columns
and one row per unique entries. Columns can be named anything you desire,
but \code{area} would be a typical entry.}

\item{collapsecharacter}{A character value that will be used to paste names
together. The default is \code{-} and it is NOT recommended to use an underscore
because underscores can cause issues in LaTeX.}
}
\value{
A vector of unique names based on row entries.
}
\description{
Get grouping structure based on a data frame of categories.
The data frame allows for any number of grouping variables,
where each row is a unique entry based on potentially multiple groups.
}
\examples{
get_groups(data.frame(group = c("North", "South")))
get_groups(data.frame(
  area = c("North", "South"),
depth = c("inshore", "offshore")
))

}
\author{
Kelli F. Johnson
}
