% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id_ling.R
\name{get_id_ling}
\alias{get_id_ling}
\title{get id for a lingcod model from the directory}
\usage{
get_id_ling(dir)
}
\arguments{
\item{dir}{model directory within lingcod package structure}
}
\description{
Uses either a \code{dir} like "models/2021.s.002.001_new_fleets"
and returns just the \code{id} part: "2021.s.002.001". \code{dir} can be a
vector of directories as well.
}
\examples{
\dontrun{
  get_id_ling(dir = "models/2021.s.002.001_new_fleets")
  # acts as the inverse of get_dir_ling()
  get_id_ling(get_dir_ling(id = "2021.s.002.001"))
}
}
\seealso{
\code{\link[=get_dir_ling]{get_dir_ling()}}
}
\author{
Ian G. Taylor
}
