% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inputs_ling.R
\name{get_inputs_ling}
\alias{get_inputs_ling}
\title{get SS input files for a lingcod model}
\usage{
get_inputs_ling(
  area = NULL,
  num = NULL,
  sens = 1,
  yr = 2021,
  id = NULL,
  dir = NULL,
  ss_new = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{area}{area, initially either "n" or "s" for 2021 models}

\item{num}{base model number}

\item{sens}{sensitivity number (assumed 1 if not specified)}

\item{yr}{model year (assumed 2021 if not specified)}

\item{id}{model id (like "2021.s.001.001") as an alternative to
the combination of inputs \code{area}, \code{num}, \code{sens}, and \code{year}}

\item{dir}{model directory (like "models/2021.s.001.001_new_fleets")
as an alternative to either \code{id} or the combination of \code{area}, \code{num},
\code{sense}, and \code{yr}}

\item{ss_new}{logical controling choice between .ss_new or standard inputs}

\item{verbose}{A logical value specifying if output should be printed
to the console or not. The default is \code{FALSE}, which will suppress messages.}
}
\description{
uses either a directory (relative to the lingcod package base directory)
or a model id like "2021.s.001.001" or the combination of
year (\code{yr}), area (\code{area}), run number (\code{num}), and sensitivity number
(\code{sens}) to assemble that id, then returns a list of the results of the
\verb{r4ss::SS_read*()} functions for the four model input files.
}
\examples{
\dontrun{
  get_inputs_ling(area ="s", num = 1)
  get_inputs_ling(id = "2021.s.001.001")
  get_inputs_ling(dir = "models/2021.s.001.001_new_fleets")
}
}
\seealso{
\code{\link[=get_dir_ling]{get_dir_ling()}}, \code{\link[=write_inputs_ling]{write_inputs_ling()}}
}
\author{
Ian G. Taylor
}
