% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dir_ling.R, R/get_mdtable.R
\name{get_mdtable}
\alias{get_mdtable}
\title{Get a markdown table from within a section in a .md file}
\usage{
get_mdtable(file, section)

get_mdtable(file, section)
}
\arguments{
\item{file}{A file path to a .md file.}

\item{section}{A text string of the section in the markdown file
that holds the table you want. This section should only contain the table,
no other text.}
}
\description{
Get a markdown table from within a section in a .md file

Get a markdown table from within a section in a .md file
}
