% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mod.R
\name{get_mod}
\alias{get_mod}
\title{get directory for a lingcod model}
\usage{
get_mod(
  area = NULL,
  num = NULL,
  sens = 1,
  yr = 2021,
  id = NULL,
  covar = TRUE,
  assign = TRUE,
  printstats = FALSE,
  plot = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{area}{area, initially either "n" or "s" for 2021 models}

\item{num}{base model number}

\item{sens}{sensitivity number (assumed 1 if not specified)}

\item{yr}{model year (assumed 2021 if not specified)}

\item{id}{model id (like "2021.s.001.001") as an alternative to
the combination of inputs \code{area}, \code{num}, \code{sens}, and \code{year}}

\item{covar}{Read covar.sso file? Passed to r4ss::SS_output() and
required in some cases where there was a non-positive-definite Hessian.}

\item{assign}{Assign the result of r4ss::SS_output() to the
user's workspace with the name \code{mod.[id]} (e.g. \code{mod.2021.s.002.001}).}

\item{printstats}{Argument passed on r4ss::SS_output()}

\item{plot}{Either TRUE/FALSE or a vector drawn from 1:26 passed to
SS_plots() to control which plot groups get created}

\item{verbose}{A logical value specifying if output should be printed
to the console or not. The default is \code{FALSE}, which will suppress messages.}

\item{dots}{Additional arguments passed to r4ss::SS_plots()}
}
\value{
invisibly returns the list created by \code{r4ss::SS_output()}
while also optionally assigning that list to the workspace
(if \code{assign = TRUE}).
}
\description{
reads model output from a directory found by \code{get_dir_ling()}
and invisibly returns the list created by \code{r4ss::SS_output()}.
}
\examples{
\dontrun{
  # read model output and assign to workspace as 'mod.2021.s.001.001"
  get_mod(area ="s", num = 1)
  get_mod(id = "2021.s.001.001")

  # read model output and make all default r4ss plots
  get_mod(area ="s", num = 2, plot = TRUE)

  # read model output and make only biology, selex, and time series plots
  get_mod(area ="s", num = 2, plot = 1:3)
}
}
\seealso{
\code{\link[=get_dir_ling]{get_dir_ling()}}, \code{\link[=get_inputs_ling]{get_inputs_ling()}}
}
\author{
Ian G. Taylor
}
