% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SS_prior.R
\name{get_SS_prior}
\alias{get_SS_prior}
\title{get prior values for prior types and parameterizations used in Stock Synthesis}
\usage{
get_SS_prior(Ptype, Pmin, Pmax, Pr, Psd, Pval)
}
\arguments{
\item{Ptype}{string explaining prior type (see code for options)}

\item{Pmin}{MIN}

\item{Pmax}{MAX}

\item{Pr}{PRIOR value (often the mean)}

\item{Psd}{PR_SD}

\item{Pval}{vector of values for which to calculate the prior}
}
\description{
copied from r4ss::SSplotPars()
https://github.com/r4ss/r4ss/blob/main/R/SSplotPars.R
can be put into r4ss as an independent function in the future
}
