% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_year.R, R/models_add.R
\name{get_year}
\alias{get_year}
\title{Get the assessment year}
\usage{
get_year(incrementafter = "September")

get_year(incrementafter = "September")
}
\arguments{
\item{incrementafter}{A character or numeric value representing the last
month you would be working on an assessment for the current year.
The default, \code{"September"}, leads to the returned value being the next year
if the current month is October or later.
If using the character version, it must be the full month name,
though case is ignored.}
}
\value{
An integer object representing the current year.

An integer object representing the current year.
}
\description{
Get the assessment year based on the current month.
If the current month is past June, then it is more than likely that you are working
on an assessment for next year and thus the assessment year would be current year + 1.

Get the assessment year based on the current month.
If the current month is past June, then it is more than likely that you are working
on an assessment for next year and thus the assessment year would be current year + 1.
}
\examples{
# Use an integer to represent the month
get_year(2)
# Use full month name
get_year("February")
# Case does not matter
get_year("may")

# Use an integer to represent the month
get_year(2)
# Use full month name
get_year("February")
# Case does not matter
get_year("may")

}
\author{
Kelli Faye Johnson

Kelli Faye Johnson
}
