% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave2.R
\name{ggsave2}
\alias{ggsave2}
\title{Save a \link[ggplot2:ggplot]{ggplot2::ggplot} figure with caption and alt text}
\usage{
ggsave2(..., caption, alttext, label)
}
\arguments{
\item{...}{Parameters passed to \link[ggplot2:ggsave]{ggplot2::ggsave}.}

\item{caption}{A character string providing the caption.}

\item{alttext}{A character string providing the alternative text.}

\item{label}{A character string providing the LaTeX label.}
}
\value{
The label used for the figure.
}
\description{
Save a \link[ggplot2:ggplot]{ggplot2::ggplot} figure and a csv file with additional
information, where the figure and csv are saved with the same
name just different extensions.
}
\author{
Kelli F. Johnson
}
