% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\docType{data}
\name{info_areacolors}
\alias{info_areacolors}
\title{Area colors}
\format{
A vector of character values which represent colors.
}
\usage{
info_areacolors
}
\description{
A vector of colors for each area
}
\keyword{datasets}
