% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\docType{data}
\name{info_authors}
\alias{info_authors}
\title{A list of authors used for writing the assessment(s)}
\format{
A list the same length as the number of documents you want to
produce, with a single vector per document.
}
\usage{
info_authors
}
\description{
A list of authors with one vector for each assessment associated
with this species / stock.
}
\seealso{
\link[sa4ss:write_authors]{sa4ss::write_authors}
}
\keyword{datasets}
