% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\docType{data}
\name{info_basemodels}
\alias{info_basemodels}
\title{Base model folder names}
\format{
A vector of character values.
}
\usage{
info_basemodels
}
\description{
A vector of names for the base models. These can be sourced by appending
\code{"models"} to the beginning to make a file path, e.g.,
\code{file.path("models", info_basemodels)}.
}
\keyword{datasets}
