% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\docType{data}
\name{info_bins}
\alias{info_bins}
\title{A list of bins used for composition data}
\format{
A list of length two, with information for age and length bins
in that order. List elements are integer values specifying the lower-left
edge of each bin and the last bin will therefore include everything
equal to and above the given integer.
}
\usage{
info_bins
}
\description{
A list of bins for binning composition data.
}
\keyword{datasets}
