% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\docType{data}
\name{info_groups}
\alias{info_groups}
\title{A data frame of groups used to structure output and the document}
\format{
A data frame where the column names are the group type
and the entries are character values for that given group type.
}
\usage{
info_groups
}
\description{
A data frame of grouping structures for working up the data and
writing the report. The data frame allows for the inclusion of as
many columns as you want. The lingcod 2021 assessment just used
a single column called area because there were two areas.
}
\keyword{datasets}
