% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\docType{data}
\name{info_surveynames}
\alias{info_surveynames}
\title{Names of included surveys}
\format{
A vector of character values.
}
\usage{
info_surveynames
}
\description{
A vector of names for the surveys included in the stock assessment.
}
\keyword{datasets}
