% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\docType{data}
\name{last_ssdat}
\alias{last_ssdat}
\title{Last assessment's SS dat files}
\format{
A list of lists, with one element for each model
that was included in the last assessment.
}
\usage{
last_ssdat
}
\description{
The dat files read in by \link[r4ss:SS_readdat]{r4ss::SS_readdat} from the last
assessment.
}
\keyword{datasets}
