% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\docType{data}
\name{last_ssdir}
\alias{last_ssdir}
\title{Last assessment's model directory}
\format{
A vector of directory names.
}
\usage{
last_ssdir
}
\description{
The directories for the last assessment located in the
models directory.
}
\keyword{datasets}
