% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_r4ss_plots_ling.R
\name{make_r4ss_plots_ling}
\alias{make_r4ss_plots_ling}
\title{Make r4ss plots for Lingcod assessment}
\usage{
make_r4ss_plots_ling(mod, plot = 1:26, verbose = TRUE, ...)
}
\arguments{
\item{mod}{List created by \code{\link[=get_mod]{get_mod()}} or \code{r4ss::SS_output()}}

\item{plot}{Vector of plot groups passed to \code{r4ss::SS_plots()}}

\item{verbose}{A logical value specifying if output should be printed
to the console or not. The default is \code{FALSE}, which will suppress messages.}

\item{dots}{additional arguments passed to \code{r4ss::SS_plots()}}
}
\description{
Wrapper for \code{r4ss::SS_plots()} applying default settings that work
for the fleets, colors, plot dimensions, etc.
}
\seealso{
\code{\link[=get_mod]{get_mod()}}
}
\author{
Ian G. Taylor
}
