% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_lbs2mt.R
\name{mult_lbs2mt}
\alias{mult_lbs2mt}
\title{Converter to get from pounds to metric tons}
\usage{
mult_lbs2mt()
}
\value{
A single numeric values
}
\description{
Multiply by this value to get from pounds to metric tons.
}
\examples{
# Convert 1 lb to metric tons
1 * mult_lbs2mt()

}
\author{
Kelli Faye Johnson
}
