% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot a map of the states with data using \link[ggplot2:ggplot]{ggplot2::ggplot}}
\usage{
plot_map(
  data,
  states = "california",
  yname = "total_catch_wt_kg",
  ytitle = "weight (kg)",
  yintercept = 34.5,
  alpha = 0.5
)
}
\arguments{
\item{data}{A data frame returned from \link[nwfscSurvey:PullCatch.fn]{nwfscSurvey::PullCatch.fn}.}

\item{states}{A single character value that will be used to search
for regions in \link[maps:map]{maps::map}. Regex is allowed, e.g.,
\code{"washington|oregon|california"} will return the mainland of the
US West Coast.}

\item{yname}{A single character value providing the name of the response
variable to include in the figure. Only values great than zero will be
included.}

\item{ytitle}{A single character value used to label the legend for the
y variable.}

\item{yintercept}{A numeric value for \link[ggplot2:geom_abline]{ggplot2::geom_hline}.
\code{NULL} will stop a horizontal line from being created.}

\item{alpha}{Passed to \link[ggplot2:geom_point]{ggplot2::geom_point} to control the opacity
of the points.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot a map of the data and include the outline of the desired states.
In theory, the data can be mapped on any state within the United States
using an variable in the data frame.
}
\examples{
plot_map(
  data =data.frame(
    yy = 1:10,
    Latitude_dd = 41:50,
    Longitude_dd = rep(-121, 10)
  ),
  yname = "yy",
  states = "washington|oregon|california",
)

}
\author{
Kelli F. Johnson
}
