% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_north_vs_south.R
\name{plot_north_vs_south}
\alias{plot_north_vs_south}
\title{Make a series of plots comparing the north and south models}
\usage{
plot_north_vs_south(mod.n, mod.s, subplot = 1:3)
}
\arguments{
\item{mod.n}{A north model as produced by \code{r4ss::SS_output()}
or \code{\link[=get_mod]{get_mod()}}}

\item{mod.s}{A south model as produced by \code{r4ss::SS_output()}
or \code{\link[=get_mod]{get_mod()}}}

\item{subplot}{Which plots to create (currently there are 3)}
}
\description{
Uses \code{r4ss::SSplotComparisons()} but applies defaults as
appropriate for the lingcod stocks
}
\examples{
\dontrun{
plot_north_vs_south(mod.2021.n.020.001, mod.2021.s.014.001)
}

}
