% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rec_selectivity_illustration.R
\name{plot_rec_selectivity_illustration}
\alias{plot_rec_selectivity_illustration}
\title{Make a figure to illustrate rec catch}
\usage{
plot_rec_selectivity_illustration()
}
\description{
See https://github.com/iantaylor-NOAA/Lingcod_2021/issues/20
}
