% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_retention.R, R/plot_selex.R
\name{plot_sel_ret}
\alias{plot_sel_ret}
\title{Plot time-varying selectivity or selectivity}
\usage{
plot_sel_ret(mod, fleet = 1, Factor = "Lsel", sex = 1)

plot_sel_ret(mod, fleet = 1, Factor = "Lsel", sex = 1)
}
\arguments{
\item{mod}{A model object created by \code{\link[=get_mod]{get_mod()}} or
\code{r4ss::SS_output()}}

\item{fleet}{a single fleet number}

\item{Factor}{a factor from mod$sizeselex$Factor}

\item{sex}{sex 1 for females, 2 for males}
}
\description{
Plot time-varying selectivity or selectivity

Plot time-varying selectivity or selectivity
}
\author{
Ian G. Taylor

Ian G. Taylor
}
