% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stock_boundary_map.R
\name{plot_stock_boundary_map}
\alias{plot_stock_boundary_map}
\title{Plot a map of the stock boundaries for the 2021 lingcod assessment}
\usage{
plot_stock_boundary_map(
  names = c("North", "South"),
  cols = c("#8290BB", "#BC7A8F")
)
}
\arguments{
\item{names}{A vector of names for north and south areas}

\item{cols}{A vector of colors for north and south areas. Default is
unikn::usecol(pal_unikn_pair, 16L)\link{c(2,10)} as used in data-raw/lingcod_catch.R
but could/should be updated to use values stored in the package}
}
\description{
Slightly cleaned up from Yellowtail Rockfish map created in 2017:
https://github.com/iantaylor-NOAA/YTRK_doc/blob/master/Rcode/map_showing_areas_v2.R
It writes a PNG file to figures/map_of_stock_boundaries_40-10.png.
This depends on the mapdata package but I got an
Error: 'worldHiresMapEnv' is not an exported object from 'namespace:maps'
so include require(mapdata) in the function as a quick solution.
}
\examples{
# make map using colors used in data-raw/lingcod_catch.R
plot_stock_boundary_map()
}
\seealso{
\code{\link[=plot_map]{plot_map()}}
}
\author{
Ian G. Taylor
}
