% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{plot_theme}
\alias{plot_theme}
\title{Plot themes for \link[ggplot2:ggplot]{ggplot2::ggplot} figures}
\usage{
plot_theme()
}
\value{
\link[ggplot2:ggplot]{ggplot2::ggplot} structures
}
\description{
Standardized plot themes for this package when using \link[ggplot2:ggplot]{ggplot2::ggplot}.
These themes should be used to increase consistency in figures created
for the stock assessment document.
}
\author{
Kelli F. Johnson
}
