% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_VASTresiduals.R
\name{plot_VASTresiduals}
\alias{plot_VASTresiduals}
\title{Plot VAST residuals}
\usage{
plot_VASTresiduals(working_dir)
}
\arguments{
\item{working_dir}{Where the VAST files are located.}

\item{nrow}{The number of rows.}
}
\description{
Plot VAST residuals
}
\examples{
\dontrun {
mapply(plot_VASTresiduals,
 dir("data-raw", pattern = "_gamma$", full = TRUE)
)
}
}
\author{
Kelli F. Johnson
}
