% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rec_catch_OR}
\alias{rec_catch_OR}
\title{Recreational catches for Oregon}
\format{
A data frame with two columns
\itemize{
\item Year: The year the data pertains to
\item mt: catches in metric tons (mt)
}
}
\usage{
rec_catch_OR
}
\description{
A data frame of recreational catches for Oregon
provided by Alison Whitman
}
\keyword{datasets}
