% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_goodlabel.R
\name{ref_goodlabel}
\alias{ref_goodlabel}
\title{Format a vector of labels for markdown}
\usage{
ref_goodlabel(label)
}
\arguments{
\item{label}{A vector of character values you want to format.}
}
\value{
A vector of formatted labels.
}
\description{
Format the labels to not have spaces or underscores.
Underscores are replaced with dashes.
}
\examples{
ref_goodlabel(c("Bad_label_with_underscores", "Bad label with spaces"))

}
\author{
Kelli F. Johnson
}
