% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lorenzenm.R
\name{run_lorenzenm}
\alias{run_lorenzenm}
\title{Run a model using Lorenzen natural mortality}
\usage{
run_lorenzenm(olddir, newdir, referenceage)
}
\arguments{
\item{olddir}{The directory you want to copy input files from.}

\item{newdir}{The new directory name for the Lorenzen natural mortality run.
If this directory already exists, it will be wiped clean prior to copying files.}

\item{referenceage}{The age at which defines, or grounds, the curve.
If there is a prior on natural mortality, the prior will apply to this age.}
}
\description{
Copy model files and choose a reference age for the
Lorenzen natural mortality curve.
}
