% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_regularization.R
\name{run_regularization}
\alias{run_regularization}
\title{Example script to run MCMC using R package adnuts for SS}
\usage{
run_regularization(dirbase, interactive = FALSE, verbose = FALSE)
}
\arguments{
\item{dirbase}{The directory that has the base-model files. This path
needs to be full or relative to your current location.}

\item{verbose}{A logical value specifying if information should be printed
to the screen or not.}

\item{interaction}{A logical value specifying if you want to run the
\link[shinystan:launch_shinyadmb]{shinystan::launch_shinyadmb}.}
}
\description{
Started Feb 2021 by Cole Monnahan (cole.monnahan@noaa.gov | AFSC)
Note I am using SS version 3.30.16, which is freely available
from: https://vlab.ncep.noaa.gov/web/stock-synthesis/home
Non-windows users will need to download and put that
executable in the hake folder.
}
\author{
Cole M. Monnahan, augmented by Kelli F. Johnson
}
