% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vast.R
\name{run_vast}
\alias{run_vast}
\title{Run VAST for a single survey}
\usage{
run_vast(
  data,
  strata,
  modelarea,
  survey,
  distribution,
  species = utils_name("latin"),
  dir = "data-raw"
)
}
\arguments{
\item{data}{A data frame downloaded from the data warehouse.}

\item{strata}{A data frame of stratifications for the US West Coast.}

\item{modelarea}{A character value, used to name the directory where
files are saved.}

\item{survey}{An available survey name, e.g., \code{"WCGBTS"}.}

\item{distribution}{An available distribution, e.g., \code{"gamma"}.
Please use \code{"compound"} for the compound Poisson-Gamma distribution.}

\item{species}{A character value with the latin name of the species
of interest.}

\item{dir}{The master directory where you want the resulting directory
to be saved.}
}
\description{
Run VAST for a single survey, where the input and output can
be easily accessed with tidy structure.
}
\author{
Kelli F. Johnson
}
