% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_output.R
\name{sens_clean_labels}
\alias{sens_clean_labels}
\title{create a nicer label for model output tables}
\usage{
sens_clean_labels(tab)
}
\arguments{
\item{tab}{a table with a \code{Label} column containing parameter labels}
}
\description{
adds a new column of english language labels
for parameters and quantities of interest
}
\author{
Ian G. Taylor
}
