% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_output.R
\name{sens_convert_offsets}
\alias{sens_convert_offsets}
\title{convert parameter offsets to standard space}
\usage{
sens_convert_offsets(tab)
}
\arguments{
\item{tab}{a table with a \code{Label} column containing parameter labels}
}
\description{
(not needed for lingcod since our male growth is modeled
as independent parameters)
}
