% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_output.R
\name{sens_convert_vals}
\alias{sens_convert_vals}
\title{Convert offset parameters}
\usage{
sens_convert_vals(tab)
}
\arguments{
\item{tab}{a table with a \code{Label} column containing parameter labels}
}
\description{
Convert log Q to standard space
and summary biomass into 1000s of mt
}
