% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_areastrata.R
\name{setup_areastrata}
\alias{setup_areastrata}
\title{Create a data frame of stratifications for each area}
\usage{
setup_areastrata(
  area_lat = c(`CAN-US` = 49, Conception = 34.45, `US-MX` = 32),
  strata_lat = nwfscDeltaGLM::strataLatitudes(),
  strata_depth = nwfscDeltaGLM::strataDepths()
)
}
\arguments{
\item{area_lat}{Latitudinal (decimal degree) breaks for each area,
including breaks for the north and south boundaries. For example,
a single area would have two values, the northern and southern borders.
It is helpful if you name the vector like the default vector that has
two areas separated at Point Conception.}

\item{strata_lat}{Latitudinal (decimal degree) breaks for each strata within
areas defined by \code{area_lat}. No need to repeat values in \code{area_lat}.
Typical values include the
Washington-Oregon border at the mouth of the Columbia River (i.e., 46.0),
Point Conception at 34.50, and
the management line at 40 degrees 10 minutes (40.166667).
Note that the first two values are rounded to the tenth of a degree and
only the management line is represented using six decimal places.
It is helpful if you name the vector like what is done for the default \code{strata_lat}.
The default vector is not named, but provides all of the available options
for stratification because we are limited by what is available in the
survey packages.}

\item{strata_depth}{Depth breaks that will be included for each strata by
area combination. Values are limited to those available in the
\code{SA3} data set stored in `data(SA3, package = "nwfscSurvey").}
}
\value{
A data frame with one row per stratification and the following columns:
\itemize{
\item name: a combination of the area and depth for the given stratification,
\item area: the calculated area (todo: units here) for the stratification,
\item Depth_m.1: the minimum depth of the stratification,
\item Depth_m.2: the maximum depth of the stratification,
\item Latitude_dd.1: the southern latitude of the stratification, and
\item Latitude_dd.2: the northern latitude of the stratification.
}
}
\description{
Create a data frame with one row per stratification and
stratification names based on which area they fall in.
Often, species will be broke into several stocks along a latitudinal
gradient as well as stratifications within each area.
Where stratifications are based on latitude and depth.
All three of these categories are accounted for here.
}
\examples{
# Strata for lingcod in 2021
setup_areastrata(
  area_lat = c("CAN-US" = 49.0, "40-10" = round(40 + 10/60, 6), "US-MX" = 32.0),
  strata_lat = c(46.0, 34.5),
  strata_depth = c(55, 183, 400)
)
}
\author{
Kelli F. Johnson
}
