% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_readdat.list.R
\name{SS_readdat.list}
\alias{SS_readdat.list}
\title{Read in a multiple SS dat files from directories into a list}
\usage{
SS_readdat.list(dir, pattern)
}
\arguments{
\item{dir}{Path to upper directory that stores all your model directories.
Inside \code{dir}, the function will search for each of the model folders
specified in \code{pattern}.}

\item{pattern}{A vector of character objects, where each entry is a
pattern that you want to search for in \code{dir}. It is best if each
entry returns only a single directory when searching, this will allow
the names of the output list to be 1:1 with respect to the input,
which is the search string that led to the path to the model directory.}
}
\value{
An invisible list, where each element of the list is the output
from \link[r4ss:SS_readdat]{r4ss::SS_readdat}. List elements will be named using the input to
\code{pattern}.
}
\description{
Read in however many SS dat files you want into a list based on using
a vector of patterns to search for various directories.
First the relevant directories are found, then the ss_new data files
are found, then each one is read into a list.
}
\author{
Kelli Faye Johnson
}
