% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_switchletter.R
\name{string_switchletter}
\alias{string_switchletter}
\title{Switch a single letter in a string with a new letter}
\usage{
string_switchletter(x, newletter)
}
\arguments{
\item{x}{A character string.}

\item{newletter}{A character string of any length you want, but
only the first letter will be used to replace the value in \code{x}.
Note that we do not pay attention to case here and all new letters
will be lower-case.}
}
\description{
Switch a single letter in a character string that is surrounded by
either full stops, underscores, or dashes with a new letter.
This is helpful when a file path has a single letter for area or
some other category and you want to easily switch it out with a new
letter. For example, "apple-a-sweet" can easily be changed to
"apple-b-sweet" with this function.
}
\author{
Kelli F. Johnson
}
