% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_acomps.R
\name{survey_acomps}
\alias{survey_acomps}
\title{Expand the age-composition data for a given survey}
\usage{
survey_acomps(sname, CAAL = FALSE, dir = "data-raw")
}
\arguments{
\item{sname}{A character value giving the survey name.}

\item{CAAL}{A logical if conditional-age-at-length compositions should
be returned. The default is to return marginal age compositions instead.}

\item{dir}{The upper directory where you would like the directories with
the results returned from this function to be saved. The default is to
use \verb{"data-raw". Survey-specific folders based on }dir\code{and}sname` will
be created.}
}
\value{
\itemize{
\item Ages
\enumerate{
\item csv files: unsexed and sexed comps for north and south
\item Figures: sex ratio and size frequency for north and south
}
\item Ages CAAL
\enumerate{
\item csv files: sex-specific comps for north and south
\item Figures: NA
}
}
}
\description{
Expand age-composition data, which REQUIRES there to be ages.
Default is marginal compositions but can do CAAL.
}
\author{
Brian J. Langseth
}
