% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_lcomps.R
\name{survey_lcomps}
\alias{survey_lcomps}
\title{Expand the length-composition data for a given survey}
\usage{
survey_lcomps(sname, doAgeRep = FALSE, dir = "data-raw")
}
\arguments{
\item{sname}{A character value giving the survey name.}

\item{doAgeRep}{A logical value.}

\item{dir}{The upper directory where you would like the directories with
the results returned from this function to be saved. The default is to
use \verb{"data-raw". Survey-specific folders based on }dir\code{and}sname` will
be created.}
}
\value{
\enumerate{
\item csv files: the unsexed and sexed comps for north and south
\item Figures: sex ratio and size frequency for north and south
}
}
\description{
Expand length-composition data, which REQUIRES there to be lengths.
}
\author{
Brian J. Langseth
}
