% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_compweight.R
\name{table_compweight}
\alias{table_compweight}
\title{Table of parameters}
\usage{
table_compweight(
  output,
 
    caption = "Data weightings applied to length and age compositions according to the `Francis' method.",
  label = "table-compweight-base"
)
}
\arguments{
\item{output}{A list from \link[r4ss:SS_output]{r4ss::SS_output}.}

\item{caption}{A character string for the caption.}

\item{label}{A character string with the label for the table.
No underscores allowed.}
}
\description{
Table of parameters
}
