% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_decision.R
\name{table_decision}
\alias{table_decision}
\title{Format a decision table}
\usage{
table_decision(
  ...,
  years,
  rowgroup = c("Constant", "ACL", "ACL2"),
  colgroup = c("Low", "Base", "High"),
  format = c("latex", "html"),
  caption = formals(kableExtra::kbl)$caption,
  label = formals(kableExtra::kbl)$label
)
}
\arguments{
\item{...}{Each element provided in the call to this function that is not
assigned to a particular input argument should be a vector of file paths
to the models you want to read for a given column of your decision table.
For example, if a single column has results from three models and
if your decision table has a low and high state of nature, then
there should be two vectors passed via \code{...} and each vector would contain
paths to three different models.
The order of the vectors will be the column order from left to right.
The order of the models within a vector will be the row order.}

\item{years}{A vector of years you want catches for.}

\item{rowgroup}{A vector of character strings to label the group names in
the first column that define the groups across rows.
Typically, this information is the catch-stream groups.
You can wrap lines if you include the makecell package in your sty file and
use \code{\link[kableExtra:linebreak]{kableExtra::linebreak()}} on this vector,
which translates the characters \verb{\\n} into something that LaTeX can use.}

\item{colgroup}{A vector of character strings to label the states of nature.}

\item{format}{Argument passed to \link[kableExtra:kbl]{kableExtra::kbl} allowing
for the switch between \code{"latex"} and \code{"html"}.}

\item{caption}{A character string that
will be passed to the \code{caption} parameter of \link[kableExtra:kbl]{kableExtra::kbl}.
The default value is \code{NULL}.}

\item{label}{A character string without underscores that
will be passed to the \code{label} parameter of \link[kableExtra:kbl]{kableExtra::kbl}.
The default value is \code{NULL}.}
}
\description{
Format a decision table for its inclusion in a document.
}
\details{
todo:
\itemize{
\item get lines below rowgroups, but I think this might be working in LaTeX only.
\item write checks
\item get vertical lines
}
}
\examples{
table_decision(
  list(mod.2021.n.022.001, mod.2021.n.022.001, mod.2021.n.022.001),
  list(mod.2021.n.022.001, mod.2021.n.022.001),
  list(mod.2021.n.022.001, mod.2021.n.022.001),
  years = 2021:2032
)
table_decision(
  list(mod.2021.n.023.611,mod.2021.n.023.612,mod.2021.n.023.613),
  list(mod.2021.n.023.621,mod.2021.n.023.622,mod.2021.n.023.623),
  list(mod.2021.n.023.631,mod.2021.n.023.632,mod.2021.n.023.633),
  years = 2021:2032
)
}
\author{
Ian G. Taylor, Chantel R. Wetzel, Kelli F. Johnson
}
