% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_dq.R
\name{table_dq}
\alias{table_dq}
\title{Table of likelihoods by source}
\usage{
table_dq(
  output,
 
    caption = "Estimates of reference points and management quantities and their associated upper and lower 95\\\\\% intervals.",
  label = "table-dq-base"
)
}
\arguments{
\item{output}{A list from \link[r4ss:SS_output]{r4ss::SS_output}.}

\item{caption}{A character string for the caption.}

\item{label}{A character string with the label for the table.
No underscores allowed.}
}
\description{
Table of likelihoods by source
}
\examples{
\dontrun{
table_likelihoods(model)
}
}
