% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_likelihoods.R
\name{table_likelihoods}
\alias{table_likelihoods}
\title{Table of likelihoods by source}
\usage{
table_likelihoods(
  output,
  caption = "Likelihoods components by source.",
  label = "table-likelihoods-base"
)
}
\arguments{
\item{output}{A list from \link[r4ss:SS_output]{r4ss::SS_output}.}

\item{caption}{A character string for the caption.}

\item{label}{A character string with the label for the table.
No underscores allowed.}
}
\description{
Table of likelihoods by source
}
\examples{
\dontrun{
table_likelihoods(model)
}
}
