% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_modelconfig.R
\name{table_modelconfig}
\alias{table_modelconfig}
\title{Summarize the configuration of the SS model}
\usage{
table_modelconfig(output)
}
\arguments{
\item{output}{A list from \link[r4ss:SS_output]{r4ss::SS_output}.}
}
\description{
Summarize the configuration of the SS model
}
\examples{
table_modelconfig(model)
}
