% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_outputconfig.R
\name{table_outputconfig}
\alias{table_outputconfig}
\title{Summarize the configuration of the SS output}
\usage{
table_outputconfig(output)
}
\arguments{
\item{output}{A list from \link[r4ss:SS_output]{r4ss::SS_output}.}
}
\description{
Summarize the configuration of the SS output
}
\examples{
table_outputconfig(output)
}
