% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_pars.R
\name{table_pars}
\alias{table_pars}
\title{Table of parameters}
\usage{
table_pars(
  output,
 
    caption = "Parameter estimates, estimation phase, parameter bounds, estimation status, estimated standard deviation (SD), prior information [distribution(mean, SD)] used in the base model.",
  label = "table-pars-base"
)
}
\arguments{
\item{output}{A list from \link[r4ss:SS_output]{r4ss::SS_output}.}

\item{caption}{A character string for the caption.}

\item{label}{A character string with the label for the table.
No underscores allowed.}
}
\description{
Table of parameters
}
