% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_profile.R
\name{table_profile}
\alias{table_profile}
\title{Summarize profile results of parameters}
\usage{
table_profile(summarylist, pargrep)
}
\arguments{
\item{summarylist}{A list of summarized Report.sso files.}

\item{pargrep}{A character value to search for using \link{grepl} on the column
\code{summarylist[["Label"]]}.}
}
\description{
Summarize profile results of parameters
}
\examples{
\dontrun{
# Read in models
profilemodels <- SSgetoutput(
  dirvec = mydir,
  keyvec = 1:((dir(mydir,pattern="^Report") \%>\% length)-1)
)
# Summarize output
profilesummary <- SSsummarize(profilemodels)
table_profile(profilesummary, pargrep = "NatM") \%>\%
knitr::kable(
  format = "latex",
  align = "r",
  position = "H",
  digits = 3,
  booktabs = TRUE,
  longtable = FALSE,
  caption = "Parameter estimates of natural mortality, $M$, from the likelihood profile over $M_{female}$ showing the correlation between the two parameters.",
  label = "profile-M",
  escape = FALSE,
  linesep = ""
) \%>\%
kableExtra::save_kable(
  file = file.path(profilemodels$replist1$inputs$dir, "profile-M.tex")
)
}
}
\author{
Kelli F. Johnson
}
