% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_projections.R
\name{table_projections}
\alias{table_projections}
\title{Table of projections}
\usage{
table_projections(
  output,
  caption = paste("Projections of potential overfishing limits (OFLs; mt),",
    "allowable biological catches (ABCs; mt),", "annual catch limits (ACLs; mt),",
    "estimated summary biomass (mt), spawning biomass (mt), and fraction unfished.",
    "Values are based on removals for the first two years.",
    "ABCs include a buffer for scientific uncertainty based on",
    "a $P^*$ of 0.45 and the category 2 default sigma = 1.0.",
    "ACLs additionally include the 40:10 adjustment for projections",
    "which fall below the $B_{40}$ reference point."),
  label = "table-projections-base"
)
}
\arguments{
\item{output}{A list from \link[r4ss:SS_output]{r4ss::SS_output}.}

\item{caption}{A character string for the caption.}

\item{label}{A character string with the label for the table.
No underscores allowed.}
}
\description{
Table of projections
}
\examples{
\dontrun{
  table_projections(model)
}
}
