% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_sens.R
\name{table_sens}
\alias{table_sens}
\title{Summarize the configuration of the SS output}
\usage{
table_sens(
  file_csv,
 
    caption = "Differences in likelihood, estimates of key parameters, and estimates of derived quantities between the base model and several alternative models (columns). Red values indicate negative log likelihoods that were lower than that for the base model.",
  dir = file.path("..", "tables")
)
}
\arguments{
\item{file_csv}{A file path to the csv file.}

\item{caption}{Text you want in the caption.}
}
\description{
Summarize the configuration of the SS output
}
\examples{
\dontrun{
table_sens("tables/sens_table_s_bio_rec.csv")
}

}
