% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_ts.R
\name{table_ts}
\alias{table_ts}
\title{Table of estimated time series}
\usage{
table_ts(
  output,
  caption = "Time series of population estimates for the base model.",
  label = "table-ts-base"
)
}
\arguments{
\item{output}{A list from \link[r4ss:SS_output]{r4ss::SS_output}.}

\item{caption}{A character string for the caption.}

\item{label}{A character string with the label for the table.
No underscores allowed.}
}
\description{
Table of estimated time series
}
\examples{
\dontrun{
table_pars(model)
}
}
