% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_fleetabb.R
\name{use_fleetabb}
\alias{use_fleetabb}
\title{Get fleet abbreviation based on gear group}
\usage{
use_fleetabb(x, pattern = "^TW|^DRG|NET", use = c("TW", "FG"))
}
\arguments{
\item{x}{A vector of character values.}

\item{pattern}{A regular expression to search for.}

\item{use}{A vector of two character values, where the first value will
be used for all values in \code{x} that match \code{pattern} and the second will
be used for those that do not match. The default are for trawl and
fixed gear abbreviations.}
}
\value{
A vector of \code{"TW"} or \code{"FG"} depending on if the values
in \code{x} match the pattern, respectively. If you change the default
values in \code{use}, then the output will be different but still one
value for TRUE and another for all FALSE \link{grepl}.
}
\description{
Get the fleet, i.e., \code{"TW"} or \code{"FG"}, depending on matching the pattern
to the values in \code{x}.
}
\examples{
use_fleet(c("HKL", "TWL", "POT"))
}
\author{
Kelli Faye Johnson
}
