% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-names.R
\name{utils_name}
\alias{utils_name}
\title{Get name}
\usage{
utils_name(
  type = c("common", "Common", "latin", "PACFIN", "area", "Area", "latitude",
    "latitude2")
)
}
\arguments{
\item{type}{The type of name that you want.
Only one type is allowed, where the first type listed will be the default.
Capitalization of the input argument will dictate how the output if formatted.
For example \code{common} leads to \code{lingcod} and \code{Common} leads to \code{Lingcod} unless
the capitalization is the same whether or not the result is a title or not
such as for latin names.}
}
\value{
A string of text based on what type of output you specify.
}
\description{
Get a name for the analysis.
}
\examples{
utils_name(type = "common")

}
\author{
Kelli Faye Johnson
}
