% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_selex.R, R/write_custom_plots_csv.R
\name{write_custom_plots_csv}
\alias{write_custom_plots_csv}
\title{Write a CSV file for custom plots}
\usage{
write_custom_plots_csv(mod, filename, caption)

write_custom_plots_csv(mod, filename, caption)
}
\arguments{
\item{mod}{A model object created by \code{\link[=get_mod]{get_mod()}} or
\code{r4ss::SS_output()}}

\item{filename}{file name for png file}

\item{caption}{caption for csv}
}
\description{
Write a CSV file for custom plots

Write a CSV file for custom plots
}
\author{
Ian G. Taylor

Ian G. Taylor
}
