% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_draft.R
\name{write_draft}
\alias{write_draft}
\title{Write the draft assessment files for this species}
\usage{
write_draft(authors, dir)
}
\arguments{
\item{authors}{A vector of author names.}

\item{dir}{A directory where you want the results saved to.
The directory must be a full path or relative to your current
working directory. Regardless, of which method you choose, the
creation of the directories is recursive and you can nest
directories as many layers as you want.}
}
\description{
Write draft assessment files for this species in separate
directories based on the input argument \code{dir}.
}
\seealso{
Other write: 
\code{\link{write_introduction}()}
}
\author{
Kelli F. Johnson
}
\concept{write}
