% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_inputs_ling.R
\name{write_inputs_ling}
\alias{write_inputs_ling}
\title{write SS input files for a lingcod model}
\usage{
write_inputs_ling(
  inputs,
  dir = NULL,
  files = c("dat", "ctl", "start", "fore"),
  datname = "ling_data.ss",
  ctlname = "ling_control.ss",
  startname = "starter.ss",
  forename = "forecast.ss",
  overwrite = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{inputs}{list created by \code{\link[=get_inputs_ling]{get_inputs_ling()}}}

\item{dir}{model directory (like "models/2021.s.001.001_new_fleets").
The \code{inputs} list includes a \code{dir} element, but this will match the
source directory prior to any changes to the model files so
requiring a new \code{dir} input will help users (or at least Ian)
ovoid accidentally overwriting the source files.}

\item{files}{vector of file types to write (default is four standard
input files--.par file not yet supported)}

\item{datname}{name of data file to write}

\item{ctlname}{name of control file to write}

\item{startname}{name of starter file to write}

\item{forename}{name of forecast file to write}
}
\description{
writes SS input files using the list created by \code{\link[=get_inputs_ling]{get_inputs_ling()}}
(presumably after modification of one or more elements)
using the \verb{r4ss::SS_write*()} functions for the four model input files.
}
\examples{
\dontrun{
  # read inputs
  inputs <- get_inputs_ling(area ="s", num = 99)
  # add new data
  inputs_with_x_changes <- inputs
  inputs_with_x_changes$dat <- add_data(inputs$dat, area = "s")
  # write the files back to the source directory
  write_inputs_ling(inputs = inputs_with_x_changes,
                    dir = "models/2021.s.099.001_with_x_changes",
                    files = "start")

  # read inputs
  inputs <- get_inputs_ling(area ="s", num = 99)
  # make small change (e.g. use the .par file)
  inputs$start$init_values_src <- 1
  # write the files back to the source directory
  write_inputs_ling(inputs = inputs,
                    dir = inputs$dir,
                    files = "start")
}
}
\seealso{
\code{\link[=get_inputs_ling]{get_inputs_ling()}},
}
\author{
Ian G. Taylor
}
