% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_introduction.R
\name{write_introduction}
\alias{write_introduction}
\title{Write the introduction file}
\usage{
write_introduction(dir, dirmaster = "..")
}
\arguments{
\item{dir}{A directory where the draft files are located.}

\item{dirmaster}{The directory, relative to \code{dir}, for the master files.
Again, this must be a relative path!
The default value assumes that master files for the actual Introduction
will be located up one directory from \code{dir}. For example, all
master files for lingcod are in a directory called doc and
the pdf is built for each area within folders called
doc\North and doc\South, which would be the values passed to dir.
If you want to use a different directory, then change the default value.}
}
\description{
Write \verb{11introduction.Rmd} to overwrite the default file
provided by \link[sa4ss:draft]{sa4ss::draft}.
}
\seealso{
Other write: 
\code{\link{write_draft}()}
}
\author{
Kelli F. Johnson
}
\concept{write}
