% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data}
\alias{add_data}
\title{add data to lingcod model}
\usage{
add_data(
  id = NULL,
  dat_type = c("catch", "index", "discard", "lencomp", "ageerror", "agecomp"),
  fleets = NULL,
  ss_new = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{id}{model id (like "2021.s.001.001") as an alternative to
the combination of inputs \code{area}, \code{num}, \code{sens}, and \code{year}}

\item{dat_type}{character vector listing data types to add}

\item{fleets}{optional vector of fleet numbers for which to add data
NULL value will use all fleets from \code{\link[=get_fleet]{get_fleet()}}}

\item{ss_new}{use data.ss_new instead of ling_data.ss}

\item{verbose}{A logical value specifying if output should be printed
to the console or not. The default is \code{FALSE}, which will suppress messages.}

\item{dir}{directory where model is located}
}
\description{
reads the data file and adds data to the model specified by
either a model \code{id} like "2021.s.002.001" or a \code{dir} where
the model file is located like "models/2021.s.002.001_new_fleets"
where the \code{id} is passed to
\code{\link[=get_inputs_ling]{get_inputs_ling()}} and from there to \code{\link[=get_dir_ling]{get_dir_ling()}} to get the
directory.
}
\seealso{
\code{\link[=get_dir_ling]{get_dir_ling()}}, \code{\link[=get_inputs_ling]{get_inputs_ling()}}
}
\author{
Ian G. Taylor
}
