% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data}
\alias{add_data}
\title{add data to lingcod model}
\usage{
add_data(
  id = "2021.s.002.001",
  dat_type = c("catch", "index", "discard", "lencomp", "ageerror", "agecomp"),
  fleets = NULL,
  ss_new = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{id}{model id (like "2021.s.002.001") for the original model
to which data will be added (replacing what's currently there for each
specified combination of data type and fleet}

\item{dat_type}{character vector listing data types to add}

\item{fleets}{optional vector of fleet numbers for which to add data
NULL value will use all fleets from \code{\link[=get_fleet]{get_fleet()}}}

\item{ss_new}{use data.ss_new instead of ling_data.ss}

\item{verbose}{logical controlling feedback about internal results}
}
\description{
adds data a model with
}
\author{
Ian G. Taylor
}
