% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data}
\alias{add_data}
\title{add data to lingcod model}
\usage{
add_data(
  dat,
  area,
  dat_type = c("catch", "index", "discard", "lencomp", "ageerror", "CAAL", "agecomp"),
  fleets = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dat}{list created by \code{\link[=get_inputs_ling]{get_inputs_ling()}} or \code{r4ss::SS_readdat()}}

\item{area}{area associated with the model, either "n" or "s"}

\item{dat_type}{character vector listing data types to add}

\item{fleets}{optional vector of fleet numbers for which to add data
NULL value will use all fleets from \code{\link[=get_fleet]{get_fleet()}}}

\item{verbose}{A logical value specifying if output should be printed
to the console or not. The default is \code{FALSE}, which will suppress messages.}
}
\description{
modifies the elements of the data file to add new data
}
\seealso{
\code{\link[=get_dir_ling]{get_dir_ling()}}, \code{\link[=get_inputs_ling]{get_inputs_ling()}}
}
\author{
Ian G. Taylor
}
