% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_retention.R, R/plot_selex.R
\name{plot_sel_comm}
\alias{plot_sel_comm}
\title{Plot selectivity and retention for the commercial fleets}
\usage{
plot_sel_comm(mod)

plot_sel_comm(mod)
}
\arguments{
\item{mod}{A model object created by \code{\link[=get_mod]{get_mod()}} or
\code{r4ss::SS_output()}}
}
\description{
Plot selectivity and retention for the commercial fleets

Plot selectivity and retention for the commercial fleets
}
\author{
Ian G. Taylor

Ian G. Taylor
}
