% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_twopanel_comparison.R
\name{plot_twopanel_comparison}
\alias{plot_twopanel_comparison}
\title{Plot a two-panel comparison using SSplotComparisons}
\usage{
plot_twopanel_comparison(
  mods,
  legendlabels = NULL,
  filename = NULL,
  print = TRUE,
  dir = "figures/compare",
  hessian = TRUE,
  endyrvec = 2021,
  ylimAdj1 = 1.05,
  ylimAdj2 = 1.05,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{mods}{A list of model objects as created by r4ss::SS_output(),
grouped together as a list(), or by r4ss::SSgetoutput()}

\item{legendlabels}{Labels for use in the legend. NULL value will
use the lingcod model id}

\item{filename}{file to write to (relative to the "figures" directory)
NULL value will use a filename like \verb{compare_[id1]_[id2].png}}

\item{print}{TRUE/FALSE print to a png file: \code{filename}
or the current graphics device}

\item{dir}{Where to put the PNG file if \code{print = TRUE}.}

\item{endyrvec}{final year to include in the figure passed to
r4ss::SSplotComparisons()}

\item{ylimAdj1}{adjustment to y-axis limit for the first plot,
relative to highest point among all models shown}

\item{ylimAdj2}{adjustment to y-axis limit for the second plot,
relative to highest point among all models shown}

\item{verbose}{A logical value specifying if output should be printed
to the console or not. The default is \code{FALSE}, which will suppress messages.}

\item{dots}{additional arguments that will get passed to
r4ss::SSplotComparisons()}
}
\description{
Plot a two-panel time series comparison of
spawning biomass and relative spawning biomass for Lingcod
}
\examples{
\dontrun{
  plot_twopanel_comparison(mods = list(mod.2019.s.002.002,
                                       mod.2019.s.003.001),
                           endyrvec = 2019)
}
}
\author{
Ian G. Taylor
}
