% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_investigatemodel.R
\name{run_investigatemodel}
\alias{run_investigatemodel}
\title{Run all model investigations for a base model}
\usage{
run_investigatemodel(
  basemodelname,
  pars = system.file("extdata", "diagpars.csv", package = "lingcod"),
  run = c("jitter", "profile", "retro")
)
}
\arguments{
\item{basemodelname}{The name of the base model of interest in \code{modeldir}.
This can be a vector of names if you want to do the same thing for each
area.}

\item{pars}{A file path the csv with the parameter specifications.
The file has a column for area, with values matching what you use in your
model name for area, e.g., 2009.n.001.001_test, would be a northern model.
and the remaining columns are parameter inputs for
\link[nwfscDiag:get_settings_profile]{nwfscDiag::get_settings_profile}, which are grabbed with \link{do.call}.}

\item{run}{A vector of options you want to run,
where the default is to run them all.}
}
\description{
Run all model investigations for a base model
}
\details{
notes:
\itemize{
\item The whole process will fail if a single model doesn't converge in the retro.
\item The process will in retro if no forecasts.
\item Error from SS_doretro I think has file names pasted together without spaces.
\verb{Error in r4ss::SS_doRetro(masterdir = retro_dir, overwrite = model_settings$overwrite,  :}
\verb{error copying file(s): ss.exeforecast.ssling_control.ssling_data.ss}
}
}
\examples{
\dontrun{
# Just need to define xxx
setwd(dirname(system.file(package = "lingcod")))
run_investigatemodel(basemodelname = basemodelname)
}
}
\author{
Chantel R. Wetzel from the example in \pkg{nwfscDiags}.
}
