% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sensitivities.R
\name{run_sensitivities}
\alias{run_sensitivities}
\title{Run sensitivities for a base model}
\usage{
run_sensitivities(
  dirbase,
  type = c("profile", "retro", "regularization", "jitter")
)
}
\arguments{
\item{dirbase}{The directory that has the base-model files. This path
needs to be full or relative to your current location.}

\item{type}{The type of sensitivity you want to run.
See the formals for options.}
}
\description{
Run sensitivities for a base model
}
\examples{
\dontrun{
dirbase <- get_dir_ling(area = area, num = 14, sens = 1)
run_sensitivities(dirbase, c("profile", "retro", "regularization"))
}
}
