% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_output.R
\name{sens_make_table}
\alias{sens_make_table}
\title{Make a table of senstivity results}
\usage{
sens_make_table(area, sens_nums, sens_type = NULL, write = FALSE)
}
\arguments{
\item{area}{which area "n" or "s"}

\item{sens_nums}{a vector of values from /inst/extdata/sensitivities.csv}

\item{sens_type}{a string indicating the type which is appended to the csv
file containing the sensitivity results}

\item{write}{logical to write csv file to doc or not}
}
\description{
Works with \code{\link[=run_sensitivities]{run_sensitivities()}} and the info in
/inst/extdata/sensitivities.csv to make a table of results
}
\author{
Ian G. Taylor
}
