% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-knitr.R
\name{utils_knit_opts}
\alias{utils_knit_opts}
\title{Set up knitr options}
\usage{
utils_knit_opts(type = c("data-raw"))
}
\arguments{
\item{type}{A character entry specifying which type of file you are rendering.
If rendering a file in}
}
\value{
Nothing is returned from this function.
}
\description{
Set up the knitr options for the rendering of various file types within
the package.
}
\author{
Kelli Faye Johnson
}
