% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_custom_plots_csv.R
\name{write_custom_plots_csv}
\alias{write_custom_plots_csv}
\title{Write a CSV file for custom plots}
\usage{
write_custom_plots_csv(mod, filepath = NULL, filename, caption)
}
\arguments{
\item{mod}{A model object created by \code{\link[=get_mod]{get_mod()}} or
\code{r4ss::SS_output()} which is used to get the directory}

\item{filepath}{An alternative to mod as a way to get the path
where the csv file be written}

\item{filename}{File name for png file}

\item{caption}{Caption for csv}
}
\description{
Write a CSV file for custom plots
}
\author{
Ian G. Taylor
}
